requirements.txt
requests==2.32.3
pyyaml==6.0.2
rich==14.0.0
ollama==0.4.8
# Optional: only needed if using backend: llamacpp-python
# pip install llama-cpp-python
